+++ R/dfts.R
@@ -226,7 +226,7 @@ dfts <- function(X, wl = 512, wl.freq = 512, length.out = 20, wn = "hanning", ov
     r <- read_wave(X = X, index = i)
     
     if (frange.dtc){
-      frng <- frd_wrblr_int(wave = r, wl = wl.freq, fsmooth = fsmooth, threshold = threshold.freq, wn = wn, flim = c(0, 22), bp = b/ 1000, ovlp = ovlp)
+      frng <- frd_wrblr_int(wave = r, wl = wl.freq, fsmooth = fsmooth, threshold = threshold.freq, wn = wn, bp = b/ 1000, ovlp = ovlp)
     
     if (!all(is.na(frng$frange))) b <- as.numeric(frng$frange) * 1000 
     }